+++ test_e2e/test-dest.sh
@@ -15,9 +15,7 @@ cd "$SCRIPT_DIR"/.. || exit 1
 export OMP_THREAD_LIMIT=1
 
 all_tests=(
-  'astradb.sh'
   'azure.sh'
-  'azure-cognitive-search.sh'
   'box.sh'
   'chroma.sh'
   #  'clarifai.sh'
@@ -28,14 +26,13 @@ all_tests=(
   'kafka-local.sh'
   'mongodb.sh'
   'opensearch.sh'
-  'pinecone.sh'
   's3.sh'
   's3_no_access.sh'
   'sharepoint-embed-cog-index.sh'
   'vectara.sh'
   'weaviate.sh'
   'milvus.sh'
-  'kdbai.sh'
+  #  'kdbai.sh'
 )
 
 full_python_matrix_tests=(